# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/posts": {
      "get": {
        "summary": "Get Posts",
        "description": "Retrieve all posts.",
        "operationId": "get_posts_posts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "type": "array",
                  "title": "Response Get Posts Posts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Post",
        "description": "Create a new post with a server-generated postId and timestamp. Accepts a JPEG or PNG image.",
        "operationId": "create_post_posts_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_post_posts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}": {
      "get": {
        "summary": "Get Post",
        "description": "Retrieve a single post by its ID.",
        "operationId": "get_post_posts__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Post",
        "description": "Update caption and/or image of a post. Supply only the fields to change.",
        "operationId": "update_post_posts__post_id__put",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_post_posts__post_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Post",
        "description": "Delete a post by its ID and remove the associated image from disk.",
        "operationId": "delete_post_posts__post_id__delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/likes": {
      "put": {
        "summary": "Increment Likes",
        "description": "Increment the like count of a post by 1.",
        "operationId": "increment_likes_posts__post_id__likes_put",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Decrement Likes",
        "description": "Decrement the like count of a post by 1, floored at 0.",
        "operationId": "decrement_likes_posts__post_id__likes_delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/comments": {
      "post": {
        "summary": "Add Comment",
        "description": "Add a new comment to a post and return the created comment.",
        "operationId": "add_comment_posts__post_id__comments_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/comments/{comment_id}": {
      "delete": {
        "summary": "Delete Comment",
        "description": "Remove a comment from a post by commentId.",
        "operationId": "delete_comment_posts__post_id__comments__comment_id__delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/comments/{comment_id}/likes": {
      "put": {
        "summary": "Like Comment",
        "description": "Increment the like count on a comment by 1.",
        "operationId": "like_comment_posts__post_id__comments__comment_id__likes_put",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Unlike Comment",
        "description": "Decrement the like count on a comment by 1, floored at 0.",
        "operationId": "unlike_comment_posts__post_id__comments__comment_id__likes_delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_post_posts_post": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "userProfilePictureUrl": {
            "type": "string",
            "title": "Userprofilepictureurl"
          },
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "username",
          "userProfilePictureUrl",
          "caption",
          "image"
        ],
        "title": "Body_create_post_posts_post"
      },
      "Body_update_post_posts__post_id__put": {
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "image": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          }
        },
        "type": "object",
        "title": "Body_update_post_posts__post_id__put"
      },
      "Comment": {
        "properties": {
          "commentId": {
            "type": "string",
            "title": "Commentid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "userProfilePictureUrl": {
            "type": "string",
            "title": "Userprofilepictureurl"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "likes": {
            "type": "integer",
            "title": "Likes",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "commentId",
          "userId",
          "username",
          "userProfilePictureUrl",
          "text",
          "timestamp"
        ],
        "title": "Comment"
      },
      "CommentCreate": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "userProfilePictureUrl": {
            "type": "string",
            "title": "Userprofilepictureurl"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "username",
          "userProfilePictureUrl",
          "text"
        ],
        "title": "CommentCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Post": {
        "properties": {
          "postId": {
            "type": "string",
            "title": "Postid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "userProfilePictureUrl": {
            "type": "string",
            "title": "Userprofilepictureurl"
          },
          "imgUrl": {
            "type": "string",
            "title": "Imgurl"
          },
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "likes": {
            "type": "integer",
            "title": "Likes",
            "default": 0
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "type": "array",
            "title": "Comments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "postId",
          "userId",
          "username",
          "userProfilePictureUrl",
          "imgUrl",
          "caption",
          "timestamp"
        ],
        "title": "Post"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}